internals: migrate the u256 from include to internals
Now that cargo-rbmt is helping ensure that bitcoin-internal types do not
leak out of public APIs, we can define the u256 type in internals. This
allows it to take advantage of all the standard cargo stuff, with
minimal risk it will accidentally be depended on by consumers.

While the u256 type's implementation remains largely the same, a few
tweaks are required.

* Visibility is now public, instead of private or public(crate), since
  the type is now exported out of internals instead of copy/pasted in.
* A simple `u256::new()` constructor is necessery now also due to the
  visibility change.
* Sprinkled in some extra docs and clippy concerns (like `#[must_use]`
  on `internals::U256::inverse()`) also triggered by the new visibility.

The large tweak is consolidation the hex implemenations. The u256 version
in bitcion used the hex-conservative dependence. To avoid adding this
dep to internals, the hand rolled unit's version is moved over instead.

A `U256Ext` trait is added to units to keep a small amount of
units-specfific logic defined in units.
